refactor: resolve compiler and cppcheck warnings across codebase

+++ include/whisperEngine.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <stdbool.h>
+#include <stddef.h>
 
 bool whisperInit(const char *modelPath, bool *use_gpu);
-bool whisperProcess(float *pcmf32, int n_samples, char *outputText, int outputLength);
+bool whisperProcess(float *pcmf32, int n_samples, char *outputText, size_t outputLength);
 void whisperFree(void);
